refactor: #49 layoutのディレクトリ変更
